refactors article.js to rely on axios baseURL
