+++ src/modules/article.js
@@ -3,18 +3,11 @@ import store from "../state/store/configureStore";
 
 const Article = {
   async index(categoryName) {
+    const options = categoryName ? { category_name: categoryName } : {};
     try {
-      let response;
-      if (categoryName) {
-        response = await axios.get(
-          "https://news-in-progress-api.herokuapp.com/api/articles/",
-          { params: { category_name: categoryName } }
-        );
-      } else {
-        response = await axios.get(
-          "https://news-in-progress-api.herokuapp.com/api/articles/"
-        );
-      }
+      let response = await axios.get("/api/articles/", {
+        params: options,
+      });
       store.dispatch({
         type: "SET_ARTICLES_INDEX",
         payload: response.data.articles,
@@ -25,9 +18,7 @@ const Article = {
   },
   async show(articleId) {
     try {
-      const response = await axios.get(
-        `https://news-in-progress-api.herokuapp.com/api/articles/${articleId}`
-      );
+      const response = await axios.get(`/api/articles/${articleId}`);
       store.dispatch({
         type: "SHOW_ARTICLE",
         payload: response.data.article,
